Hack together a basic scrolling neck background.
The code is horrid, as is most of the code in game.cpp/hpp but that whole setup is going to be replaced once we get a transform hierarchy, and better object managment.

The neck works by scrolling texture coordinates on a plane that moves with the camera.

To clean this up I might want to look into ways of declairing shader uniforms per batch?